# Improve performance of Arr::dot method - 300x in some cases